+++ src/postgres/9-updates.sql
@@ -79,5 +79,18 @@ RAISE;
     UPDATE input SET enums='["Minute(s)","Hour(s)","Day(s)","Week(s)","Month(s)","Year(s)","Instantaneous","Varies"]' where id='91c123bf-702e-458c-90a1-b26f6498937e'
     UPDATE question SET long_name='Ongoing Data Production', text='After this data product has been published at the DAAC, will you continue to collect or produce new data files to extend the temporal coverage?' where id='dbf70495-433a-439e-8173-0cb20f972c16';
     UPDATE question SET help='The temporal coverage should encompass the beginning date of the first data file and the ending date of the last data file at the time of initial delivery to the DAAC, even if there are time gaps or data production will be ongoing.' where id='4f2dd369-d865-47ba-8504-8694493f129f';
-
     UPDATE question SET long_name='Data Product Spatial Region' where id='228cb0d6-78fb-449a-8061-b1e6fb3f59d1';
+
+-- 7/12/2023 Added department to contact information
+    UPDATE input set list_order=4 where question_id = '80ac5f52-9ed9-4139-b5f9-7b4cebb6a8e2' and control_id='data_producer_info_orcid';
+    UPDATE input set list_order=3 where question_id = '80ac5f52-9ed9-4139-b5f9-7b4cebb6a8e2' and control_id='data_producer_info_email';
+    INSERT INTO input VALUES ('80ac5f52-9ed9-4139-b5f9-7b4cebb6a8e2', 'data_producer_info_department', 2, 'Department', 'text', '{}', '{}', '[]','[]',  False);
+    
+    UPDATE input set list_order=4 where question_id = 'f3e2eab9-6375-4e53-9cc2-3d16f318d333' and control_id='poc_orcid';
+    UPDATE input set list_order=3 where question_id = 'f3e2eab9-6375-4e53-9cc2-3d16f318d333' and control_id='poc_email';
+    INSERT INTO input VALUES ('f3e2eab9-6375-4e53-9cc2-3d16f318d333', 'poc_department', 2, 'Department', 'text', '{}', '{}', '[]','[]',  False);
+
+    UPDATE input set list_order=4 where question_id = 'f3e2eab9-6375-4e53-9cc2-3d16f318d332' and control_id='long_term_support_poc_orcid';
+    UPDATE input set list_order=3 where question_id = 'f3e2eab9-6375-4e53-9cc2-3d16f318d332' and control_id='long_term_support_poc_email';
+    INSERT INTO input VALUES ('f3e2eab9-6375-4e53-9cc2-3d16f318d332', 'long_term_support_poc_department', 2, 'Department', 'text', '{}', '{}', '[]','[]',  False);
+